--- a/.gitignore
+++ b/.gitignore
@@ -42,3 +42,4 @@ cairo_venv
 # solidity cache
 starknet-e2e-test/contracts/cache
 starknet-e2e-test/contracts/build
+tmp/