--- a/crates/client/rpc/Cargo.toml
+++ b/crates/client/rpc/Cargo.toml
@@ -42,7 +42,7 @@ sc-network-sync = { workspace = true }
 blockifier = { workspace = true, default-features = true }
 starknet-core = { workspace = true }
 starknet-ff = { workspace = true }
-starknet_api = { workspace = true, default-features = false }
+starknet_api = { workspace = true, default-features = true }
 # Others
 anyhow = { workspace = true }
 hex = { workspace = true, default-features = true }