--- a/crates/primitives/block/src/lib.rs
+++ b/crates/primitives/block/src/lib.rs
@@ -1,5 +1,4 @@
 //! Starknet block primitives.
-#![cfg_attr(not(feature = "std"), no_std)]
 
 #[doc(hidden)]
 pub extern crate alloc;