--- a/crates/primitives/chain-id/src/lib.rs
+++ b/crates/primitives/chain-id/src/lib.rs
@@ -1,5 +1,4 @@
 //! Traits for chain ids.
-#![cfg_attr(not(feature = "std"), no_std)]
 
 use mp_felt::Felt252Wrapper;