--- a/crates/primitives/fee/src/lib.rs
+++ b/crates/primitives/fee/src/lib.rs
@@ -1,5 +1,4 @@
 //! Starknet fee logic
-#![cfg_attr(not(feature = "std"), no_std)]
 
 #[doc(hidden)]
 pub extern crate alloc;