--- a/crates/primitives/hashers/src/lib.rs
+++ b/crates/primitives/hashers/src/lib.rs
@@ -1,5 +1,4 @@
 //! Traits for hashing.
-#![cfg_attr(not(feature = "std"), no_std)]
 
 #[doc(hidden)]
 pub extern crate alloc;