--- a/crates/primitives/sequencer-address/src/lib.rs
+++ b/crates/primitives/sequencer-address/src/lib.rs
@@ -1,5 +1,4 @@
 //! The address of the account receiving the network fee
-#![cfg_attr(not(feature = "std"), no_std)]
 
 use sp_inherents::{InherentData, InherentIdentifier, IsFatalError};
 use thiserror_no_std::Error;