--- a/crates/primitives/simulations/src/lib.rs
+++ b/crates/primitives/simulations/src/lib.rs
@@ -1,4 +1,3 @@
-#![cfg_attr(not(feature = "std"), no_std)]
 
 #[doc(hidden)]
 pub extern crate alloc;