--- a/crates/primitives/snos-output/src/lib.rs
+++ b/crates/primitives/snos-output/src/lib.rs
@@ -1,5 +1,4 @@
 //! StarkNet OS program output primitives.
-#![cfg_attr(not(feature = "std"), no_std)]
 
 #[doc(hidden)]
 pub extern crate alloc;