--- a/crates/primitives/state/src/lib.rs
+++ b/crates/primitives/state/src/lib.rs
@@ -1,5 +1,4 @@
 //! Starknet state primitives.
-#![cfg_attr(not(feature = "std"), no_std)]
 
 use blockifier::execution::contract_class::ContractClass;
 use blockifier::state::cached_state::{ContractStorageKey, StateChangesCount};