--- a/crates/primitives/state/src/rpc.rs
+++ b/crates/primitives/state/src/rpc.rs
@@ -1,5 +1,4 @@
 //! Starknet rpc state primitives.
-#![cfg_attr(not(feature = "std"), no_std)]
 
 #[doc(hidden)]
 pub extern crate alloc;