--- a/crates/primitives/storage/src/lib.rs
+++ b/crates/primitives/storage/src/lib.rs
@@ -1,5 +1,4 @@
 //! Starknet storage primitives.
-#![cfg_attr(not(feature = "std"), no_std)]
 
 extern crate alloc;
 use alloc::vec::Vec;