--- a/crates/primitives/transactions/src/lib.rs
+++ b/crates/primitives/transactions/src/lib.rs
@@ -1,5 +1,4 @@
 //! Starknet transaction related functionality.
-#![cfg_attr(not(feature = "std"), no_std)]
 
 #[doc(hidden)]
 pub extern crate alloc;