--- a/crates/runtime/build.rs
+++ b/crates/runtime/build.rs
@@ -1,5 +1,5 @@
-use substrate_wasm_builder::WasmBuilder;
+//use substrate_wasm_builder::WasmBuilder;
 
 fn main() {
-    WasmBuilder::new().with_current_project().export_heap_base().import_memory().build()
+    //WasmBuilder::new().with_current_project().export_heap_base().import_memory().build()
 }