--- a/crates/runtime/src/lib.rs
+++ b/crates/runtime/src/lib.rs
@@ -8,7 +8,8 @@
 
 // Make the WASM binary available.
 #[cfg(feature = "std")]
-include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
+//include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
+pub const WASM_BINARY: Option<&[u8]> = Some(&[]);
 
 /// Runtime modules.
 mod config;
@@ -456,7 +457,7 @@ impl_runtime_apis! {
         fn dispatch_benchmark(
             config: frame_benchmarking::BenchmarkConfig
         ) -> Result<Vec<frame_benchmarking::BenchmarkBatch>, sp_runtime::RuntimeString> {
-            use frame_benchmarking::{baseline, Benchmarking, BenchmarkBatch, TrackedStorageKey};
+            use frame_benchmarking::{baseline, Benchmarking, BenchmarkBatch};
 
             use frame_system_benchmarking::Pallet as SystemBench;
             use baseline::Pallet as BaselineBench;
@@ -465,7 +466,7 @@ impl_runtime_apis! {
             impl baseline::Config for Runtime {}
 
             use frame_support::traits::WhitelistedStorageKeys;
-            let whitelist: Vec<TrackedStorageKey> = AllPalletsWithSystem::whitelisted_storage_keys();
+            let whitelist: Vec<_> = AllPalletsWithSystem::whitelisted_storage_keys();
 
             let mut batches = Vec::<BenchmarkBatch>::new();
             let params = (&config, &whitelist);