### PR goal

Explore and find the pain points to make Madara std-compatible.

### Main tasks

Exploration

- Explore Madara github to check which dependencies are no_std
- Explore Substrate / Polkadot-sdk github to understand the impact of the changes

### Exploration

Various solutions have been considered:

- Fork substrate to force the usage of the NativeExecutor
- Write host functions for features that need STD, that would be called by the WasmExecutor
- Try and revert the changes made in substrate (the changes that aim to make the WasmExecutor the only way to execute the runtime)

## Findings

The first solution seems to be the easiest to implement, as the current substrate code still give access to the NativeExecutor.

### Blocking points

- When enabling the “std” feature, the build fails when trying to build the Wasm Runtime, as it tries to include std features in a target (wasm32-unknown-unknown) that does not support them.

### Changes implemented

- Fork the [blockifier](https://github.com/massalabs/blockifier/tree/no_std-support-7578442-std) to target our fork of polkadot-sdk and enable all the default-features (”std”)
    - Patches [c66b1..](https://github.com/massalabs/blockifier/commit/c66b1eb5acd49a69c10fe908a101a0b6a9ed53c4) and [b9e80..](https://github.com/massalabs/blockifier/commit/b9e8048d7bae683cad7164de277716de27a8c3db)
- [Fork madara](https://github.com/massalabs/madara-std) to:
    - Target [our fork of polkadot-sdk](https://github.com/massalabs/polkadot-sdk/tree/release-polkadot-v1.3.0-std)
        - Patch [7b7ac2..](https://github.com/massalabs/madara-std/pull/1/commits/7b7ac2fd14d05efff2414a3da69db869c4baa43e)
    - Target [our fork of the blockifier](https://github.com/massalabs/blockifier/tree/no_std-support-7578442-std)
        - Patch [7b7ac2..](https://github.com/massalabs/madara-std/pull/1/commits/7b7ac2fd14d05efff2414a3da69db869c4baa43e)
    - Enable the default-features (”std”) for all madara’s dependancies.
        - Patch [c7ebb80..](https://github.com/massalabs/madara-std/pull/1/commits/c7ebb80f9abb9c5af764940310d4caf9fa9acbd1)
    - Disable the build of the Wasm runtime (through substrate’s WasmBuilder)
        - Patch [7b7ac2..](https://github.com/massalabs/madara-std/pull/1/commits/7b7ac2fd14d05efff2414a3da69db869c4baa43e)
    - Remove the no_std code
        - Patch [70ada09..](https://github.com/massalabs/madara-std/pull/1/commits/70ada0906b8d3c11d37df9f51ef8471d406fe02d)
    - Expose a dummy empty `WASM_BINARY` constant to allow instantiation of the NativeElseWasmExecutor
        - Patch [7b7ac2..](https://github.com/massalabs/madara-std/pull/1/commits/7b7ac2fd14d05efff2414a3da69db869c4baa43e)
- Fork [polkadot-sdk](https://github.com/massalabs/polkadot-sdk/tree/release-polkadot-v1.3.0-std) to:
    - The Wasm runtime was called in the `runtime_version` and `call` methods on the NativeElseWasmExecutor. This was blocking because even if the Native executor could handle the call, it would still call the Wasm runtime. So when calling the runtime in substrate, we force the use of the Native runtime and don’t instantiate / call the Wasm runtime (which is now empty).
        - Patch [a41673..](https://github.com/massalabs/polkadot-sdk/commit/a41673f8aa3f24d7ed8fb1f246a97551eca58837)

### Results

With the changes mentioned above:

- The project successfully builds
- All `cargo test` passes
- The chain runs
    
    ```bash
    cargo run --release -- setup --chain=dev --from-remote
    cargo run --release -- --dev
    ```
    
- The block benchmark runs
    
    ```bash
    cargo run --release -- setup --base-path ./tmp --chain=dev --from-remote
    cargo run --release -- --base-path ./tmp --chain=dev --force-authoring --alice --rpc-external --rpc-methods=unsafe
    cargo run --release -- benchmark block --from 1 --to 3 --base-path ./tmp --chain=dev
    ```
    
- Benchmark results:

|  | baseline (with no_std) | with our changes |
| --- | --- | --- |
| Mean block time | 2,678,720 ns | 581,360 ns |
| Mean block weight usage | 280% | 60.9 % |
| Block weight | 954,238 ns | 954,238 ns |

This means that using the native executor leads to a significant increase in performance.

## Notes

- Using the NativeExecutor (instead of the WasmExecutor) disable the ability to update the chain without interruption.
- Here is the [PR on the Blockifier's side](https://github.com/keep-starknet-strange/blockifier/pull/27)