@Leo-Besancon 2024-02-02
@tdelabro seems like the linter is broken on main so it also fails here.