@tdelabro 2024-02-02
@Leo-Besancon I'm trying to fix it with https://github.com/keep-starknet-strange/madara/pull/1429